fix and simplify tests
